Refactor project index resource analysis
